Refactoring posizione files
