+++ src/app.module.ts
@@ -4,8 +4,8 @@ import { RouteReuseStrategy } from '@angular/router';
 
 import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
 
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app/modules/app-routing.module';
+import { AppComponent } from './app/components/app/app.component';
 
 @NgModule({
   declarations: [AppComponent],
