+++ src/app/components/explore-container/explore-container.component.scss
