+++ src/app/components/tab1/tab1.module.ts
