+++ src/app/components/tabs/tabs.module.ts
