+++ src/app/components/tabs/tabs.page.scss
