+++ src/app/components/tabs/tabs.page.ts
